refactor: move institution API type ownership
