+++ frontend/src/api/institutions/types.ts
@@ -1,4 +1,11 @@
-export type { Institution } from '@/api/accounts';
+export interface Institution {
+  id: string;
+  status: string;
+  name: string;
+  country_code: string;
+  website: string;
+  logo_url: string | null;
+}
 
 export interface CreateInstitutionPayload {
   name: string;
